+++ src/routes/auth.routes.js
@@ -1,12 +1,14 @@
 import { Router } from "express";
-import {signup, signin, logout} from "../controllers/auth.controller.js";
+import { signup } from "../controllers/registro.controller.js";
+import { signin } from "../controllers/logeo.controller.js";
+import { logout } from "../controllers/deslogeo.controller.js";
 
 const router = Router();
 
-router.post("/auth/registrarse", signup);
+router.post("/alterlearn/registrarse", signup);
 
-router.post("/auth/logearse", signin);
+router.post("/alterlearn/logearse", signin);
 
-router.get("/auth/deslogearse", logout);
+router.get("/alterlearn/deslogearse", logout);
 
 export default router;
